Router: ignore line breaks in def/val assignments

+++ repos/spark/mllib/src/main/scala/org/apache/spark/ml/regression/GeneralizedLinearRegression.scala
@@ -56,13 +56,14 @@ private[regression] trait GeneralizedLinearRegressionBase
     * @group param
     */
   @Since("2.0.0")
-  final val family: Param[String] = new Param(
-    this,
-    "family",
-    "The name of family which is a description of the error distribution to be used in the " +
-      "model. Supported options: gaussian(default), binomial, poisson and gamma.",
-    ParamValidators.inArray[String](
-      GeneralizedLinearRegression.supportedFamilyNames.toArray))
+  final val family: Param[String] =
+    new Param(
+      this,
+      "family",
+      "The name of family which is a description of the error distribution to be used in the " +
+        "model. Supported options: gaussian(default), binomial, poisson and gamma.",
+      ParamValidators.inArray[String](
+        GeneralizedLinearRegression.supportedFamilyNames.toArray))
 
   /** @group getParam */
   @Since("2.0.0")
@@ -75,15 +76,16 @@ private[regression] trait GeneralizedLinearRegressionBase
     * @group param
     */
   @Since("2.0.0")
-  final val link: Param[String] = new Param(
-    this,
-    "link",
-    "The name of link function " +
-      "which provides the relationship between the linear predictor and the mean of the " +
-      "distribution function. Supported options: identity, log, inverse, logit, probit, " +
-      "cloglog and sqrt.",
-    ParamValidators.inArray[String](
-      GeneralizedLinearRegression.supportedLinkNames.toArray))
+  final val link: Param[String] =
+    new Param(
+      this,
+      "link",
+      "The name of link function " +
+        "which provides the relationship between the linear predictor and the mean of the " +
+        "distribution function. Supported options: identity, log, inverse, logit, probit, " +
+        "cloglog and sqrt.",
+      ParamValidators.inArray[String](
+        GeneralizedLinearRegression.supportedLinkNames.toArray))
 
   /** @group getParam */
   @Since("2.0.0")
@@ -215,11 +217,12 @@ class GeneralizedLinearRegression @Since("2.0.0") (
   override protected def train(
       dataset: DataFrame): GeneralizedLinearRegressionModel = {
     val familyObj = Family.fromName($(family))
-    val linkObj = if (isDefined(link)) {
-      Link.fromName($(link))
-    } else {
-      familyObj.defaultLink
-    }
+    val linkObj =
+      if (isDefined(link)) {
+        Link.fromName($(link))
+      } else {
+        familyObj.defaultLink
+      }
     val familyAndLink = new FamilyAndLink(familyObj, linkObj)
 
     val numFeatures = dataset
@@ -253,11 +256,12 @@ class GeneralizedLinearRegression @Since("2.0.0") (
 
     if (familyObj == Gaussian && linkObj == Identity) {
       // TODO: Make standardizeFeatures and standardizeLabel configurable.
-      val optimizer = new WeightedLeastSquares(
-        $(fitIntercept),
-        $(regParam),
-        standardizeFeatures = true,
-        standardizeLabel = true)
+      val optimizer =
+        new WeightedLeastSquares(
+          $(fitIntercept),
+          $(regParam),
+          standardizeFeatures = true,
+          standardizeLabel = true)
       val wlsModel = optimizer.fit(instances)
       val model = copyValues(
         new GeneralizedLinearRegressionModel(
@@ -266,27 +270,31 @@ class GeneralizedLinearRegression @Since("2.0.0") (
           wlsModel.intercept)
           .setParent(this))
       // Handle possible missing or invalid prediction columns
-      val (summaryModel, predictionColName) =
-        model.findSummaryModelAndPredictionCol()
-      val trainingSummary = new GeneralizedLinearRegressionSummary(
-        summaryModel.transform(dataset),
-        predictionColName,
-        model,
-        wlsModel.diagInvAtWA.toArray,
-        1)
+      val (summaryModel, predictionColName) = model
+        .findSummaryModelAndPredictionCol()
+      val trainingSummary =
+        new GeneralizedLinearRegressionSummary(
+          summaryModel.transform(dataset),
+          predictionColName,
+          model,
+          wlsModel.diagInvAtWA.toArray,
+          1)
       return model.setSummary(trainingSummary)
     }
 
     // Fit Generalized Linear Model by iteratively reweighted least squares (IRLS).
-    val initialModel =
-      familyAndLink.initialize(instances, $(fitIntercept), $(regParam))
-    val optimizer = new IterativelyReweightedLeastSquares(
-      initialModel,
-      familyAndLink.reweightFunc,
+    val initialModel = familyAndLink.initialize(
+      instances,
       $(fitIntercept),
-      $(regParam),
-      $(maxIter),
-      $(tol))
+      $(regParam))
+    val optimizer =
+      new IterativelyReweightedLeastSquares(
+        initialModel,
+        familyAndLink.reweightFunc,
+        $(fitIntercept),
+        $(regParam),
+        $(maxIter),
+        $(tol))
     val irlsModel = optimizer.fit(instances)
 
     val model = copyValues(
@@ -296,14 +304,15 @@ class GeneralizedLinearRegression @Since("2.0.0") (
         irlsModel.intercept)
         .setParent(this))
     // Handle possible missing or invalid prediction columns
-    val (summaryModel, predictionColName) =
-      model.findSummaryModelAndPredictionCol()
-    val trainingSummary = new GeneralizedLinearRegressionSummary(
-      summaryModel.transform(dataset),
-      predictionColName,
-      model,
-      irlsModel.diagInvAtWA.toArray,
-      irlsModel.numIterations)
+    val (summaryModel, predictionColName) = model
+      .findSummaryModelAndPredictionCol()
+    val trainingSummary =
+      new GeneralizedLinearRegressionSummary(
+        summaryModel.transform(dataset),
+        predictionColName,
+        model,
+        irlsModel.diagInvAtWA.toArray,
+        irlsModel.numIterations)
 
     model.setSummary(trainingSummary)
   }
@@ -338,12 +347,12 @@ object GeneralizedLinearRegression
   )
 
   /** Set of family names that GeneralizedLinearRegression supports. */
-  private[ml] lazy val supportedFamilyNames =
-    supportedFamilyAndLinkPairs.map(_._1.name)
+  private[ml] lazy val supportedFamilyNames = supportedFamilyAndLinkPairs.map(
+    _._1.name)
 
   /** Set of link names that GeneralizedLinearRegression supports. */
-  private[ml] lazy val supportedLinkNames =
-    supportedFamilyAndLinkPairs.map(_._2.name)
+  private[ml] lazy val supportedLinkNames = supportedFamilyAndLinkPairs.map(
+    _._2.name)
 
   private[ml] val epsilon: Double = 1e-16
 
@@ -372,12 +381,13 @@ object GeneralizedLinearRegression
         Instance(eta, instance.weight, instance.features)
       }
       // TODO: Make standardizeFeatures and standardizeLabel configurable.
-      val initialModel = new WeightedLeastSquares(
-        fitIntercept,
-        regParam,
-        standardizeFeatures = true,
-        standardizeLabel = true)
-        .fit(newInstances)
+      val initialModel =
+        new WeightedLeastSquares(
+          fitIntercept,
+          regParam,
+          standardizeFeatures = true,
+          standardizeLabel = true)
+          .fit(newInstances)
       initialModel
     }
 
@@ -754,11 +764,12 @@ class GeneralizedLinearRegressionModel private[ml] (
   import GeneralizedLinearRegression._
 
   lazy val familyObj = Family.fromName($(family))
-  lazy val linkObj = if (isDefined(link)) {
-    Link.fromName($(link))
-  } else {
-    familyObj.defaultLink
-  }
+  lazy val linkObj =
+    if (isDefined(link)) {
+      Link.fromName($(link))
+    } else {
+      familyObj.defaultLink
+    }
   lazy val familyAndLink = new FamilyAndLink(familyObj, linkObj)
 
   override protected def predict(features: Vector): Double = {
@@ -868,10 +879,11 @@ object GeneralizedLinearRegressionModel
       val intercept = data.getDouble(0)
       val coefficients = data.getAs[Vector](1)
 
-      val model = new GeneralizedLinearRegressionModel(
-        metadata.uid,
-        coefficients,
-        intercept)
+      val model =
+        new GeneralizedLinearRegressionModel(
+          metadata.uid,
+          coefficients,
+          intercept)
 
       DefaultParamsReader.getAndSetParams(model, metadata)
       model
@@ -902,22 +914,24 @@ class GeneralizedLinearRegressionSummary private[regression] (
   import GeneralizedLinearRegression._
 
   private lazy val family = Family.fromName(model.getFamily)
-  private lazy val link = if (model.isDefined(model.getParam("link"))) {
-    Link.fromName(model.getLink)
-  } else {
-    family.defaultLink
-  }
+  private lazy val link =
+    if (model.isDefined(model.getParam("link"))) {
+      Link.fromName(model.getLink)
+    } else {
+      family.defaultLink
+    }
 
   /** Number of instances in DataFrame predictions */
   private lazy val numInstances: Long = predictions.count()
 
   /** The numeric rank of the fitted linear model */
   @Since("2.0.0")
-  lazy val rank: Long = if (model.getFitIntercept) {
-    model.coefficients.size + 1
-  } else {
-    model.coefficients.size
-  }
+  lazy val rank: Long =
+    if (model.getFitIntercept) {
+      model.coefficients.size + 1
+    } else {
+      model.coefficients.size
+    }
 
   /** Degrees of freedom */
   @Since("2.0.0")
@@ -931,11 +945,12 @@ class GeneralizedLinearRegressionSummary private[regression] (
 
   /** The residual degrees of freedom for the null model */
   @Since("2.0.0")
-  lazy val residualDegreeOfFreedomNull: Long = if (model.getFitIntercept) {
-    numInstances - 1
-  } else {
-    numInstances
-  }
+  lazy val residualDegreeOfFreedomNull: Long =
+    if (model.getFitIntercept) {
+      numInstances - 1
+    } else {
+      numInstances
+    }
 
   private lazy val devianceResiduals: DataFrame = {
     val drUDF = udf { (y: Double, mu: Double, weight: Double) =>
@@ -1019,13 +1034,15 @@ class GeneralizedLinearRegressionSummary private[regression] (
         lit(1.0)
       else
         col(model.getWeightCol)
-    val wtdmu: Double = if (model.getFitIntercept) {
-      val agg =
-        predictions.agg(sum(w.multiply(col(model.getLabelCol))), sum(w)).first()
-      agg.getDouble(0) / agg.getDouble(1)
-    } else {
-      link.unlink(0.0)
-    }
+    val wtdmu: Double =
+      if (model.getFitIntercept) {
+        val agg = predictions
+          .agg(sum(w.multiply(col(model.getLabelCol))), sum(w))
+          .first()
+        agg.getDouble(0) / agg.getDouble(1)
+      } else {
+        link.unlink(0.0)
+      }
     predictions
       .select(col(model.getLabelCol), w)
       .rdd
@@ -1106,11 +1123,12 @@ class GeneralizedLinearRegressionSummary private[regression] (
     */
   @Since("2.0.0")
   lazy val tValues: Array[Double] = {
-    val estimate = if (model.getFitIntercept) {
-      Array.concat(model.coefficients.toArray, Array(model.intercept))
-    } else {
-      model.coefficients.toArray
-    }
+    val estimate =
+      if (model.getFitIntercept) {
+        Array.concat(model.coefficients.toArray, Array(model.intercept))
+      } else {
+        model.coefficients.toArray
+      }
     estimate.zip(coefficientStandardErrors).map { x =>
       x._1 / x._2
     }
